requirements.txt
Django~=4.0.5
pip~=22.1.1
screen~=1.0.1
attrs~=21.4.0
toml~=0.10.2
cryptography==36.0.2
requests~=2.27.1
py~=1.11.0
Pillow==9.0.1
Jinja2==3.1.2
setuptools==62.3.2
drf-yasg==1.20.0
django-environ==0.8.1
django-filter~=21.1
gcc7==0.0.6
psycopg2-binary==2.9.3
djangorestframework~=3.13.1
pyparsing==3.0.5
coreschema
python-dotenv~=0.19.2
social-auth-app-django==5.0.0
social-auth-app-django-mongoengine==1.0.0

mongoengine==0.24.1
# environ~=1.0

environ~=1.0
pydantic~=1.9.1
pytest~=7.1.2
Faker~=11.3.0
freezegun~=1.2.1